@@ -1,18 +1,25 @@
 name: jukevault
 description: Flutter Plugin used to query audios/songs infos [title, artist, album, etc..] from device storage.
-version: 0.0.1
+version: 0.0.2
 publish_to: 'none' 
 
 environment:
-  # TODO: Update
-  sdk: ">=2.12.0 <3.0.0"
-  flutter: ">=1.20.0"
+  sdk: ">=3.0.0 <4.0.0"
+  flutter: ">=3.0.0 <4.0.0"
 
 dependencies:
   jukevault_platform_interface:
-    path: ../jukevault_platform_interface
+    # path: ../jukevault_platform_interface
+    git:
+      url: https://github.com/igraysoni/jukevault
+      path: packages/jukevault_platform_interface
+      ref: main
   jukevault_android:
-    path: ../jukevault_android
+    # path: ../jukevault_android
+    git:
+      url: https://github.com/igraysoni/jukevault
+      path: packages/jukevault_android
+      ref: main
 
   # Flutter
   flutter:
@@ -22,7 +29,7 @@ dependencies:
 
 dev_dependencies:
   # Flutter
-  flutter_lints: ^2.0.1
+  flutter_lints: ^4.0.0
   flutter_test:
     sdk: flutter
 
