@@ -1,16 +1,19 @@
 name: jukevault_android
 description: Android implementation of the jukevault plugin.
-version: 0.0.1
+version: 0.0.2
 publish_to: 'none' 
 
 environment:
-  #TODO: update version
-  sdk: ">=2.17.0 <3.0.0"
-  flutter: ">=2.5.0"
+  sdk: ">=3.0.0 <4.0.0"
+  flutter: ">=3.0.0 <4.0.0"
 
 dependencies:
   jukevault_platform_interface:
-    path: ../jukevault_platform_interface
+    # path: ../jukevault_platform_interface
+    git:
+      url: https://github.com/igraysoni/jukevault
+      path: packages/jukevault_platform_interface
+      ref: main
 
   # Flutter
   flutter:
@@ -18,7 +21,7 @@ dependencies:
 
 dev_dependencies:
   # Flutter
-  flutter_lints: ^2.0.1
+  flutter_lints: ^4.0.0
   flutter_test:
     sdk: flutter
 
