@@ -1,20 +1,19 @@
 name: jukevault_platform_interface
 description: A common platform interface for the [jukevault] plugin.
-version: 0.0.1
+version: 0.0.2
 publish_to: 'none'
 
 environment:
-  #TODO: Up SDK and Flutter versions
-  sdk: ">=2.12.0 <3.0.0"
-  flutter: ">=2.0.0"
+  sdk: ">=3.0.0 <4.0.0"
+  flutter: ">=3.0.0 <4.0.0"
 
 dependencies:
-  plugin_platform_interface: ^2.1.3
+  plugin_platform_interface: ^2.1.8
 
   # TODO: Discontinued, find a new package
   id3: ^1.0.2
-  path: ^1.8.2
-  path_provider: ^2.0.12
+  path: ^1.9.0
+  path_provider: ^2.1.4
 
   # Flutter
   flutter:
@@ -22,6 +21,6 @@ dependencies:
 
 dev_dependencies:
   # Flutter
-  flutter_lints: ^2.0.1
+  flutter_lints: ^4.0.0
   flutter_test:
     sdk: flutter
