@@ -1,6 +1,6 @@
 // ignore_for_file: constant_identifier_names
 
-part of sorts_controller;
+part of '../../controllers/sorts_controller.dart';
 
 /// Defines the sort type used for [querySongs] and [queryAudios].
 enum AudioSortType {
